Implement the same thing as the last commit.
- BC functions now follow the same pattern

- BC user functions now implement UserFun trait
