+++ src/bin/sbl.rs
@@ -6,6 +6,7 @@ use sbl::common::*;
 use sbl::errors::*;
 use sbl::vm::*;
 use sbl::compile::*;
+use sbl::internal::*;
 use sbl::bc::*;
 use std::process;
 use std::env;
@@ -34,7 +35,7 @@ fn run_program<P: AsRef<Path>, Q: AsRef<Path>>(
     };
     if dump {
         for f in fun_table.iter().filter_map(
-            |(_, f)| if let &BCFun::UserFun(ref f) =
+            |(_, f)| if let &Fun::UserFun(ref f) =
                 f as &BCFun
             {
                 Some(f)
